Extract the logic into a function under utils and document
