+++ frontend/src/index.jsx
@@ -5,7 +5,7 @@ import { Provider } from 'react-redux';
 import App from './containers/App';
 import configureStore from './store/configureStore';
 import { ProvideAuth } from './hooks/useAuth';
-import { updateLastBackPage } from './actions/AppActions';
+import { historyDoubleBackOnPopstate } from './utils';
 
 const store = configureStore();
 
@@ -44,16 +44,5 @@ ReactDOM.render(
 // to deal with this case we added a `popstate` listener that will go to the correct page in history skipping
 // the case when the URL and the view are the same when the back button is pressed
 window.addEventListener('popstate', () => {
-  const appState = store.getState().appHandler;
-  const { lastBackPage } = appState;
-
-  if (
-    lastBackPage &&
-    lastBackPage.includes('viewId') &&
-    lastBackPage === document.location.href
-  ) {
-    window.history.back(-2);
-  }
-
-  store.dispatch(updateLastBackPage(document.location.href));
+  historyDoubleBackOnPopstate(store);
 });
